refactor: extract environment
